requirements.txt
# AutoApply Bot Dependencies
# Install with: pip install -r requirements.txt

# Web scraping
playwright==1.43.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.2.1

# Resume parsing
pdfplumber==0.11.0

# AI matching
# anthropic==0.28.0   # only needed if using Anthropic
openai==1.30.0        # using OpenAI for matching

# Excel tracking
openpyxl==3.1.2

# Scheduling
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.1
